requirements.txt
Django==4.2.16
reportlab==4.2.2
Pillow==10.4.0
num2words==0.5.13
gunicorn==22.0.0
whitenoise==6.7.0
psycopg2-binary==2.9.10
python-decouple==3.8
dj-database-url==2.2.0
requests==2.31.0
